create separate backends for plotly and matplotlib
