+++ nilearn/plotting/surface/surf_plotting.py
@@ -1,24 +1,23 @@
 """Functions for surface visualization."""
 
 import itertools
-import math
-from collections.abc import Sequence
 from warnings import warn
 
-import matplotlib as mpl
+from matplotlib import __version__ as mpl_version
 import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
 from matplotlib import gridspec
-from matplotlib.cm import ScalarMappable
-from matplotlib.colorbar import make_axes
 from matplotlib.colors import LinearSegmentedColormap, Normalize, to_rgba
 from matplotlib.patches import Patch
 from mpl_toolkits.mplot3d.art3d import Poly3DCollection
 
 from nilearn import DEFAULT_DIVERGING_CMAP, image, surface
 from nilearn._utils import check_niimg_3d, compare_version, fill_doc
-from nilearn._utils.helpers import is_kaleido_installed, is_plotly_installed
+from nilearn._utils.helpers import (
+    is_matplotlib_installed,
+    is_plotly_installed
+)
 from nilearn._utils.param_validation import check_params
 from nilearn.plotting._utils import (
     check_surface_plotting_inputs,
@@ -28,10 +27,11 @@ from nilearn.plotting._utils import (
     sanitize_hemi_for_surface_image,
     save_figure_if_needed,
 )
-from nilearn.plotting.cm import mix_colormaps
 from nilearn.plotting.displays import PlotlySurfaceFigure
-from nilearn.plotting.html_surface import get_vertexcolor
-from nilearn.plotting.js_plotting_utils import colorscale
+from nilearn.plotting.surface._utils import (
+    _check_hemispheres,
+    _check_views,
+)
 from nilearn.surface import (
     load_surf_data,
     load_surf_mesh,
@@ -43,17 +43,29 @@ from nilearn.surface.surface import (
     check_mesh_is_fsaverage,
 )
 
-VALID_VIEWS = (
-    "anterior",
-    "posterior",
-    "medial",
-    "lateral",
-    "dorsal",
-    "ventral",
-    "left",
-    "right",
-)
-VALID_HEMISPHERES = "left", "right", "both"
+
+def _get_surface_backend(engine=None):
+    if engine == "matplotlib":
+        if is_matplotlib_installed():
+            from nilearn.plotting.surface._matplotlib import MatplotlibBackend
+            return MatplotlibBackend()
+        else:
+            msg = "Using engine='matplotlib' requires that ``matplotlib`` is installed."
+            raise ImportError(msg)
+    elif engine == "plotly":
+        if is_plotly_installed():
+            from nilearn.plotting.surface._plotly import PlotlyBackend
+            return PlotlyBackend()
+        else:
+            msg = "Using engine='plotly' requires that ``plotly`` is installed."
+            raise ImportError(msg)
+    else:
+        raise ValueError(
+            f"Unknown plotting engine {engine}. "
+            "Please use either 'matplotlib' or "
+            "'plotly'."
+        )
+
 
 # subset of data format extensions supported
 DATA_EXTENSIONS = (
@@ -61,446 +73,6 @@ DATA_EXTENSIONS = (
     "gii.gz",
     "mgz",
 )
-
-
-MATPLOTLIB_VIEWS = {
-    "left": {
-        "lateral": (0, 180),
-        "medial": (0, 0),
-        "dorsal": (90, 0),
-        "ventral": (270, 0),
-        "anterior": (0, 90),
-        "posterior": (0, 270),
-    },
-    "right": {
-        "lateral": (0, 0),
-        "medial": (0, 180),
-        "dorsal": (90, 0),
-        "ventral": (270, 0),
-        "anterior": (0, 90),
-        "posterior": (0, 270),
-    },
-    "both": {
-        "right": (0, 0),
-        "left": (0, 180),
-        "dorsal": (90, 0),
-        "ventral": (270, 0),
-        "anterior": (0, 90),
-        "posterior": (0, 270),
-    },
-}
-
-
-CAMERAS = {
-    "left": {
-        "eye": {"x": -1.5, "y": 0, "z": 0},
-        "up": {"x": 0, "y": 0, "z": 1},
-        "center": {"x": 0, "y": 0, "z": 0},
-    },
-    "right": {
-        "eye": {"x": 1.5, "y": 0, "z": 0},
-        "up": {"x": 0, "y": 0, "z": 1},
-        "center": {"x": 0, "y": 0, "z": 0},
-    },
-    "dorsal": {
-        "eye": {"x": 0, "y": 0, "z": 1.5},
-        "up": {"x": -1, "y": 0, "z": 0},
-        "center": {"x": 0, "y": 0, "z": 0},
-    },
-    "ventral": {
-        "eye": {"x": 0, "y": 0, "z": -1.5},
-        "up": {"x": 1, "y": 0, "z": 0},
-        "center": {"x": 0, "y": 0, "z": 0},
-    },
-    "anterior": {
-        "eye": {"x": 0, "y": 1.5, "z": 0},
-        "up": {"x": 0, "y": 0, "z": 1},
-        "center": {"x": 0, "y": 0, "z": 0},
-    },
-    "posterior": {
-        "eye": {"x": 0, "y": -1.5, "z": 0},
-        "up": {"x": 0, "y": 0, "z": 1},
-        "center": {"x": 0, "y": 0, "z": 0},
-    },
-}
-
-
-AXIS_CONFIG = {
-    "showgrid": False,
-    "showline": False,
-    "ticks": "",
-    "title": "",
-    "showticklabels": False,
-    "zeroline": False,
-    "showspikes": False,
-    "spikesides": False,
-    "showbackground": False,
-}
-
-
-LAYOUT = {
-    "scene": {
-        "dragmode": "orbit",
-        **{f"{dim}axis": AXIS_CONFIG for dim in ("x", "y", "z")},
-    },
-    "paper_bgcolor": "#fff",
-    "hovermode": False,
-    "margin": {"l": 0, "r": 0, "b": 0, "t": 0, "pad": 0},
-}
-
-
-def _get_camera_view_from_string_view(hemi, view):
-    """Return plotly camera parameters from string view."""
-    if hemi in ["left", "right"]:
-        if view == "lateral":
-            return CAMERAS[hemi]
-        elif view == "medial":
-            return CAMERAS[
-                (
-                    VALID_HEMISPHERES[0]
-                    if hemi == VALID_HEMISPHERES[1]
-                    else VALID_HEMISPHERES[1]
-                )
-            ]
-    elif hemi == "both" and view in ["lateral", "medial"]:
-        raise ValueError(
-            "Invalid view definition: when hemi is 'both', "
-            "view cannot be 'lateral' or 'medial'.\n"
-            "Maybe you meant 'left' or 'right'?"
-        )
-    return CAMERAS[view]
-
-
-def _get_camera_view_from_elevation_and_azimut(view):
-    """Compute plotly camera parameters from elevation and azimut."""
-    elev, azim = view
-    # The radius is useful only when using a "perspective" projection,
-    # otherwise, if projection is "orthographic",
-    # one should tweak the "aspectratio" to emulate zoom
-    r = 1.5
-    # The camera position and orientation is set by three 3d vectors,
-    # whose coordinates are independent of the plotted data.
-    return {
-        # Where the camera should look at
-        # (it should always be looking at the center of the scene)
-        "center": {"x": 0, "y": 0, "z": 0},
-        # Where the camera should be located
-        "eye": {
-            "x": (
-                r
-                * math.cos(azim / 360 * 2 * math.pi)
-                * math.cos(elev / 360 * 2 * math.pi)
-            ),
-            "y": (
-                r
-                * math.sin(azim / 360 * 2 * math.pi)
-                * math.cos(elev / 360 * 2 * math.pi)
-            ),
-            "z": r * math.sin(elev / 360 * 2 * math.pi),
-        },
-        # How the camera should be rotated.
-        # It is determined by a 3d vector indicating which direction
-        # should look up in the generated plot
-        "up": {
-            "x": math.sin(elev / 360 * 2 * math.pi)
-            * math.cos(azim / 360 * 2 * math.pi + math.pi),
-            "y": math.sin(elev / 360 * 2 * math.pi)
-            * math.sin(azim / 360 * 2 * math.pi + math.pi),
-            "z": math.cos(elev / 360 * 2 * math.pi),
-        },
-        # "projection": {"type": "perspective"},
-        "projection": {"type": "orthographic"},
-    }
-
-
-def _get_view_plot_surf_plotly(hemi, view):
-    """
-    Get camera parameters from hemi and view for the plotly engine.
-
-    This function checks the selected hemisphere and view, and
-    returns the cameras view.
-    """
-    _check_views([view])
-    _check_hemispheres([hemi])
-    if isinstance(view, str):
-        return _get_camera_view_from_string_view(hemi, view)
-    return _get_camera_view_from_elevation_and_azimut(view)
-
-
-def _configure_title_plotly(title, font_size, color="black"):
-    """Help for plot_surf with plotly engine.
-
-    This function configures the title if provided.
-    """
-    if title is None:
-        return {}
-    return {
-        "text": title,
-        "font": {
-            "size": font_size,
-            "color": color,
-        },
-        "y": 0.96,
-        "x": 0.5,
-        "xanchor": "center",
-        "yanchor": "top",
-    }
-
-
-def _get_cbar_plotly(
-    colorscale,
-    vmin,
-    vmax,
-    cbar_tick_format,
-    fontsize=25,
-    color="black",
-    height=0.5,
-):
-    """Help for _plot_surf_plotly.
-
-    This function configures the colorbar and creates a small
-    invisible plot that uses the appropriate cmap to trigger
-    the generation of the colorbar. This dummy plot has then to
-    be added to the figure.
-    """
-    dummy = {
-        "opacity": 0,
-        "colorbar": {
-            "tickfont": {"size": fontsize, "color": color},
-            "tickformat": cbar_tick_format,
-            "len": height,
-        },
-        "type": "mesh3d",
-        "colorscale": colorscale,
-        "x": [1, 0, 0],
-        "y": [0, 1, 0],
-        "z": [0, 0, 1],
-        "i": [0],
-        "j": [1],
-        "k": [2],
-        "intensity": [0.0],
-        "cmin": vmin,
-        "cmax": vmax,
-    }
-    return dummy
-
-
-def _plot_surf_plotly(
-    coords,
-    faces,
-    surf_map=None,
-    bg_map=None,
-    hemi="left",
-    view="lateral",
-    cmap=None,
-    symmetric_cmap=True,
-    colorbar=False,
-    threshold=None,
-    bg_on_data=False,
-    darkness=0.7,
-    vmin=None,
-    vmax=None,
-    cbar_tick_format=".1f",
-    title=None,
-    title_font_size=18,
-    output_file=None,
-):
-    """Help for plot_surf.
-
-    .. versionadded:: 0.9.0
-
-    This function handles surface plotting when the selected
-    engine is plotly.
-
-    .. note::
-        This function assumes that plotly and kaleido are
-        installed.
-
-    .. warning::
-        This function is new and experimental. Please report
-        bugs that you may encounter.
-
-    """
-    if is_plotly_installed():
-        import plotly.graph_objects as go
-    else:
-        msg = "Using engine='plotly' requires that ``plotly`` is installed."
-        raise ImportError(msg)
-
-    x, y, z = coords.T
-    i, j, k = faces.T
-
-    if cmap is None:
-        cmap = DEFAULT_DIVERGING_CMAP
-
-    bg_data = None
-    if bg_map is not None:
-        bg_data = load_surf_data(bg_map)
-        if bg_data.shape[0] != coords.shape[0]:
-            raise ValueError(
-                "The bg_map does not have the same number "
-                "of vertices as the mesh."
-            )
-
-    if surf_map is not None:
-        _check_surf_map(surf_map, coords.shape[0])
-        colors = colorscale(
-            cmap,
-            surf_map,
-            threshold,
-            vmax=vmax,
-            vmin=vmin,
-            symmetric_cmap=symmetric_cmap,
-        )
-        vertexcolor = get_vertexcolor(
-            surf_map,
-            colors["cmap"],
-            colors["norm"],
-            absolute_threshold=colors["abs_threshold"],
-            bg_map=bg_data,
-            bg_on_data=bg_on_data,
-            darkness=darkness,
-        )
-    else:
-        if bg_data is None:
-            bg_data = np.zeros(coords.shape[0])
-        colors = colorscale("Greys", bg_data, symmetric_cmap=False)
-        vertexcolor = get_vertexcolor(
-            bg_data,
-            colors["cmap"],
-            colors["norm"],
-            absolute_threshold=colors["abs_threshold"],
-        )
-
-    mesh_3d = go.Mesh3d(x=x, y=y, z=z, i=i, j=j, k=k, vertexcolor=vertexcolor)
-    fig_data = [mesh_3d]
-    if colorbar:
-        dummy = _get_cbar_plotly(
-            colors["colors"],
-            float(colors["vmin"]),
-            float(colors["vmax"]),
-            cbar_tick_format,
-        )
-        fig_data.append(dummy)
-
-    # instantiate plotly figure
-    camera_view = _get_view_plot_surf_plotly(hemi, view)
-    fig = go.Figure(data=fig_data)
-    fig.update_layout(
-        scene_camera=camera_view,
-        title=_configure_title_plotly(title, title_font_size),
-        **LAYOUT,
-    )
-
-    # save figure
-    plotly_figure = PlotlySurfaceFigure(
-        figure=fig, output_file=output_file, hemi=hemi
-    )
-
-    if output_file is not None:
-        if not is_kaleido_installed():
-            msg = (
-                "Saving figures to file with engine='plotly' requires "
-                "that ``kaleido`` is installed."
-            )
-            raise ImportError(msg)
-        plotly_figure.savefig()
-
-    return plotly_figure
-
-
-def _get_view_plot_surf_matplotlib(hemi, view):
-    """Help function for plot_surf with matplotlib engine.
-
-    This function checks the selected hemisphere and view, and
-    returns elev and azim.
-    """
-    _check_views([view])
-    _check_hemispheres([hemi])
-    if isinstance(view, str):
-        if hemi == "both" and view in ["lateral", "medial"]:
-            raise ValueError(
-                "Invalid view definition: when hemi is 'both', "
-                "view cannot be 'lateral' or 'medial'.\n"
-                "Maybe you meant 'left' or 'right'?"
-            )
-        return MATPLOTLIB_VIEWS[hemi][view]
-    return view
-
-
-def _check_surf_map(surf_map, n_vertices):
-    """Help for plot_surf.
-
-    This function checks the dimensions of provided surf_map.
-    """
-    surf_map_data = load_surf_data(surf_map)
-    if surf_map_data.ndim != 1:
-        raise ValueError(
-            "'surf_map' can only have one dimension "
-            f"but has '{surf_map_data.ndim}' dimensions"
-        )
-    if surf_map_data.shape[0] != n_vertices:
-        raise ValueError(
-            "The surf_map does not have the same number "
-            "of vertices as the mesh."
-        )
-    return surf_map_data
-
-
-def _compute_surf_map_faces_matplotlib(
-    surf_map, faces, avg_method, n_vertices, face_colors_size
-):
-    """Help for plot_surf.
-
-    This function computes the surf map faces using the
-    provided averaging method.
-
-    .. note::
-        This method is called exclusively when using matplotlib,
-        since it only supports plotting face-colour maps and not
-        vertex-colour maps.
-
-    """
-    surf_map_data = _check_surf_map(surf_map, n_vertices)
-
-    # create face values from vertex values by selected avg methods
-    error_message = (
-        "avg_method should be either "
-        "['mean', 'median', 'max', 'min'] "
-        "or a custom function"
-    )
-    if isinstance(avg_method, str):
-        try:
-            avg_method = getattr(np, avg_method)
-        except AttributeError:
-            raise ValueError(error_message)
-        surf_map_faces = avg_method(surf_map_data[faces], axis=1)
-    elif callable(avg_method):
-        surf_map_faces = np.apply_along_axis(
-            avg_method, 1, surf_map_data[faces]
-        )
-
-        # check that surf_map_faces has the same length as face_colors
-        if surf_map_faces.shape != (face_colors_size,):
-            raise ValueError(
-                "Array computed with the custom function "
-                "from avg_method does not have the correct shape: "
-                f"{surf_map_faces[0]} != {face_colors_size}"
-            )
-
-        # check that dtype is either int or float
-        if not (
-            "int" in str(surf_map_faces.dtype)
-            or "float" in str(surf_map_faces.dtype)
-        ):
-            raise ValueError(
-                "Array computed with the custom function "
-                "from avg_method should be an array of numbers "
-                "(int or float)"
-            )
-    else:
-        raise ValueError(error_message)
-    return surf_map_faces
 
 
 def _get_ticks_matplotlib(vmin, vmax, cbar_tick_format, threshold):
@@ -518,257 +90,6 @@ def _get_ticks_matplotlib(vmin, vmax, cbar_tick_format, threshold):
         return get_cbar_ticks(vmin, vmax, threshold, n_ticks)
 
 
-def _get_cmap_matplotlib(cmap, vmin, vmax, cbar_tick_format, threshold=None):
-    """Help for plot_surf with matplotlib engine.
-
-    This function returns the colormap.
-    """
-    our_cmap = plt.get_cmap(cmap)
-    norm = Normalize(vmin=vmin, vmax=vmax)
-    cmaplist = [our_cmap(i) for i in range(our_cmap.N)]
-    if threshold is not None:
-        if cbar_tick_format == "%i" and int(threshold) != threshold:
-            warn(
-                "You provided a non integer threshold "
-                "but configured the colorbar to use integer formatting."
-            )
-        # set colors to gray for absolute values < threshold
-        istart = int(norm(-threshold, clip=True) * (our_cmap.N - 1))
-        istop = int(norm(threshold, clip=True) * (our_cmap.N - 1))
-        for i in range(istart, istop):
-            cmaplist[i] = (0.5, 0.5, 0.5, 1.0)
-    our_cmap = LinearSegmentedColormap.from_list(
-        "Custom cmap", cmaplist, our_cmap.N
-    )
-    return our_cmap, norm
-
-
-def _compute_facecolors_matplotlib(bg_map, faces, n_vertices, darkness, alpha):
-    """Help for plot_surf with matplotlib engine.
-
-    This function computes the facecolors.
-    """
-    if bg_map is None:
-        bg_data = np.ones(n_vertices) * 0.5
-    else:
-        bg_data = np.copy(load_surf_data(bg_map))
-        if bg_data.shape[0] != n_vertices:
-            raise ValueError(
-                "The bg_map does not have the same number "
-                "of vertices as the mesh."
-            )
-
-    bg_faces = np.mean(bg_data[faces], axis=1)
-    # scale background map if need be
-    bg_vmin, bg_vmax = np.min(bg_faces), np.max(bg_faces)
-    if bg_vmin < 0 or bg_vmax > 1:
-        bg_norm = mpl.colors.Normalize(vmin=bg_vmin, vmax=bg_vmax)
-        bg_faces = bg_norm(bg_faces)
-
-    if darkness is not None:
-        bg_faces *= darkness
-        warn(
-            (
-                "The `darkness` parameter will be deprecated in release 0.13. "
-                "We recommend setting `darkness` to None"
-            ),
-            DeprecationWarning,
-        )
-
-    face_colors = plt.cm.gray_r(bg_faces)
-
-    # set alpha if in auto mode
-    if alpha == "auto":
-        alpha = 0.5 if bg_map is None else 1
-    # modify alpha values of background
-    face_colors[:, 3] = alpha * face_colors[:, 3]
-
-    return face_colors
-
-
-def _threshold_and_rescale(data, threshold, vmin, vmax):
-    """Help for plot_surf.
-
-    This function thresholds and rescales the provided data.
-    """
-    data_copy, vmin, vmax = _rescale(data, vmin, vmax)
-    return data_copy, _threshold(data, threshold, vmin, vmax), vmin, vmax
-
-
-def _threshold(data, threshold, vmin, vmax):
-    """Thresholds the data."""
-    # If no thresholding and nans, filter them out
-    if threshold is None:
-        mask = np.logical_not(np.isnan(data))
-    else:
-        mask = np.abs(data) >= threshold
-        if vmin > -threshold:
-            mask = np.logical_and(mask, data >= vmin)
-        if vmax < threshold:
-            mask = np.logical_and(mask, data <= vmax)
-    return mask
-
-
-def _rescale(data, vmin=None, vmax=None):
-    """Rescales the data."""
-    data_copy = np.copy(data)
-    # if no vmin/vmax are passed figure them out from data
-    vmin, vmax = _get_bounds(data_copy, vmin, vmax)
-    data_copy -= vmin
-    data_copy /= vmax - vmin
-    return data_copy, vmin, vmax
-
-
-def _get_bounds(data, vmin=None, vmax=None):
-    """Help returning the data bounds."""
-    vmin = np.nanmin(data) if vmin is None else vmin
-    vmax = np.nanmax(data) if vmax is None else vmax
-    return vmin, vmax
-
-
-def _plot_surf_matplotlib(
-    coords,
-    faces,
-    surf_map=None,
-    bg_map=None,
-    hemi="left",
-    view="lateral",
-    cmap=None,
-    colorbar=False,
-    avg_method="mean",
-    threshold=None,
-    alpha="auto",
-    bg_on_data=False,
-    darkness=0.7,
-    vmin=None,
-    vmax=None,
-    cbar_vmin=None,
-    cbar_vmax=None,
-    cbar_tick_format="%.2g",
-    title=None,
-    output_file=None,
-    axes=None,
-    figure=None,
-):
-    """Help for plot_surf.
-
-    This function handles surface plotting when the selected
-    engine is matplotlib.
-    """
-    _default_figsize = [4, 5]
-    limits = [coords.min(), coords.max()]
-
-    # Get elevation and azimut from view
-    elev, azim = _get_view_plot_surf_matplotlib(hemi, view)
-
-    # if no cmap is given, set to matplotlib default
-    if cmap is None:
-        cmap = plt.get_cmap(plt.rcParamsDefault["image.cmap"])
-    # if cmap is given as string, translate to matplotlib cmap
-    elif isinstance(cmap, str):
-        cmap = plt.get_cmap(cmap)
-
-    figsize = _default_figsize
-    # Leave space for colorbar
-    if colorbar:
-        figsize[0] += 0.7
-    # initiate figure and 3d axes
-    if axes is None:
-        if figure is None:
-            figure = plt.figure(figsize=figsize)
-        axes = figure.add_axes((0, 0, 1, 1), projection="3d")
-    elif figure is None:
-        figure = axes.get_figure()
-    axes.set_xlim(*limits)
-    axes.set_ylim(*limits)
-    axes.view_init(elev=elev, azim=azim)
-    axes.set_axis_off()
-
-    # plot mesh without data
-    p3dcollec = axes.plot_trisurf(
-        coords[:, 0],
-        coords[:, 1],
-        coords[:, 2],
-        triangles=faces,
-        linewidth=0.1,
-        antialiased=False,
-        color="white",
-    )
-
-    # reduce viewing distance to remove space around mesh
-    axes.set_box_aspect(None, zoom=1.3)
-
-    bg_face_colors = _compute_facecolors_matplotlib(
-        bg_map, faces, coords.shape[0], darkness, alpha
-    )
-    if surf_map is not None:
-        surf_map_faces = _compute_surf_map_faces_matplotlib(
-            surf_map,
-            faces,
-            avg_method,
-            coords.shape[0],
-            bg_face_colors.shape[0],
-        )
-        surf_map_faces, kept_indices, vmin, vmax = _threshold_and_rescale(
-            surf_map_faces, threshold, vmin, vmax
-        )
-
-        surf_map_face_colors = cmap(surf_map_faces)
-        # set transparency of voxels under threshold to 0
-        surf_map_face_colors[~kept_indices, 3] = 0
-        if bg_on_data:
-            # if need be, set transparency of voxels above threshold to 0.7
-            # so that background map becomes visible
-            surf_map_face_colors[kept_indices, 3] = 0.7
-
-        face_colors = mix_colormaps(surf_map_face_colors, bg_face_colors)
-
-        if colorbar:
-            cbar_vmin = cbar_vmin if cbar_vmin is not None else vmin
-            cbar_vmax = cbar_vmax if cbar_vmax is not None else vmax
-            ticks = _get_ticks_matplotlib(
-                cbar_vmin, cbar_vmax, cbar_tick_format, threshold
-            )
-            our_cmap, norm = _get_cmap_matplotlib(
-                cmap, vmin, vmax, cbar_tick_format, threshold
-            )
-            bounds = np.linspace(cbar_vmin, cbar_vmax, our_cmap.N)
-
-            # we need to create a proxy mappable
-            proxy_mappable = ScalarMappable(cmap=our_cmap, norm=norm)
-            proxy_mappable.set_array(surf_map_faces)
-            cax, _ = make_axes(
-                axes,
-                location="right",
-                fraction=0.15,
-                shrink=0.5,
-                pad=0.0,
-                aspect=10.0,
-            )
-            figure.colorbar(
-                proxy_mappable,
-                cax=cax,
-                ticks=ticks,
-                boundaries=bounds,
-                spacing="proportional",
-                format=cbar_tick_format,
-                orientation="vertical",
-            )
-
-        # fix floating point bug causing highest to sometimes surpass 1
-        # (for example 1.0000000000000002)
-        face_colors[face_colors > 1] = 1
-
-        p3dcollec.set_facecolors(face_colors)
-        p3dcollec.set_edgecolors(face_colors)
-
-    if title is not None:
-        axes.set_title(title)
-
-    return save_figure_if_needed(figure, output_file)
-
-
-@fill_doc
 def plot_surf(
     surf_mesh=None,
     surf_map=None,
@@ -982,97 +303,40 @@ def plot_surf(
     if view is None:
         view = "dorsal" if hemi == "both" else "lateral"
 
-    parameters_not_implemented_in_plotly = {
-        "avg_method": avg_method,
-        "figure": figure,
-        "axes": axes,
-        "cbar_vmin": cbar_vmin,
-        "cbar_vmax": cbar_vmax,
-        "alpha": alpha,
-    }
-
     surf_map, surf_mesh, bg_map = check_surface_plotting_inputs(
         surf_map, surf_mesh, hemi, bg_map
     )
 
     check_extensions(surf_map, DATA_EXTENSIONS, FREESURFER_DATA_EXTENSIONS)
 
-    if engine == "plotly":
-        for parameter, value in parameters_not_implemented_in_plotly.items():
-            if value is not None:
-                warn(
-                    f"'{parameter}' is not implemented "
-                    "for the plotly engine.\n"
-                    f"Got '{parameter} = {value}'.\n"
-                    f"Use '{parameter} = None' to silence this warning."
-                )
-
     coords, faces = load_surf_mesh(surf_mesh)
 
-    if engine == "matplotlib":
-        # setting defaults
-        if avg_method is None:
-            avg_method = "mean"
-        if alpha is None:
-            alpha = "auto"
-
-        if cbar_tick_format == "auto":
-            cbar_tick_format = "%.2g"
-        fig = _plot_surf_matplotlib(
-            coords,
-            faces,
-            surf_map=surf_map,
-            bg_map=bg_map,
-            hemi=hemi,
-            view=view,
-            cmap=cmap,
-            colorbar=colorbar,
-            avg_method=avg_method,
-            threshold=threshold,
-            alpha=alpha,
-            bg_on_data=bg_on_data,
-            darkness=darkness,
-            vmin=vmin,
-            vmax=vmax,
-            cbar_vmin=cbar_vmin,
-            cbar_vmax=cbar_vmax,
-            cbar_tick_format=cbar_tick_format,
-            title=title,
-            output_file=output_file,
-            axes=axes,
-            figure=figure,
-        )
-
-    elif engine == "plotly":
-        if cbar_tick_format == "auto":
-            cbar_tick_format = ".1f"
-        fig = _plot_surf_plotly(
-            coords,
-            faces,
-            surf_map=surf_map,
-            bg_map=bg_map,
-            view=view,
-            hemi=hemi,
-            cmap=cmap,
-            symmetric_cmap=symmetric_cmap,
-            colorbar=colorbar,
-            threshold=threshold,
-            bg_on_data=bg_on_data,
-            darkness=darkness,
-            vmin=vmin,
-            vmax=vmax,
-            cbar_tick_format=cbar_tick_format,
-            title=title,
-            title_font_size=title_font_size,
-            output_file=output_file,
-        )
-
-    else:
-        raise ValueError(
-            f"Unknown plotting engine {engine}. "
-            "Please use either 'matplotlib' or "
-            "'plotly'."
-        )
+    fig = _get_surface_backend(engine).plot_surf(
+        coords,
+        faces,
+        surf_map=surf_map,
+        bg_map=bg_map,
+        hemi=hemi,
+        view=view,
+        cmap=cmap,
+        symmetric_cmap=symmetric_cmap,
+        colorbar=colorbar,
+        avg_method=avg_method,
+        threshold=threshold,
+        alpha=alpha,
+        bg_on_data=bg_on_data,
+        darkness=darkness,
+        vmin=vmin,
+        vmax=vmax,
+        cbar_vmin=cbar_vmin,
+        cbar_vmax=cbar_vmax,
+        cbar_tick_format=cbar_tick_format,
+        title=title,
+        title_font_size=title_font_size,
+        output_file=output_file,
+        axes=axes,
+        figure=figure,
+    )
 
     return fig
 
@@ -1274,7 +538,7 @@ def plot_surf_contours(
         # Fix: Matplotlib version 3.3.2 to 3.3.3
         # Attribute _facecolors3d changed to _facecolor3d in
         # matplotlib version 3.3.3
-        if compare_version(mpl.__version__, "<", "3.3.3"):
+        if compare_version(mpl_version, "<", "3.3.3"):
             axes.collections[0]._facecolors3d[faces_outside] = color
             if axes.collections[0]._edgecolors3d.size == 0:
                 axes.collections[0].set_edgecolor(
@@ -1554,79 +818,6 @@ def plot_surf_stat_map(
         **kwargs,
     )
     return display
-
-
-def _check_hemisphere_is_valid(hemi):
-    return hemi in VALID_HEMISPHERES
-
-
-def _check_hemispheres(hemispheres):
-    """Check whether the hemispheres passed to in plot_img_on_surf are \
-    correct.
-
-    hemispheres : :obj:`list`
-        Any combination of 'left' and 'right'.
-
-    """
-    invalid_hemis = [
-        not _check_hemisphere_is_valid(hemi) for hemi in hemispheres
-    ]
-    if any(invalid_hemis):
-        raise ValueError(
-            "Invalid hemispheres definition!\n"
-            f"Got: {np.array(hemispheres)[invalid_hemis]!s}\n"
-            f"Supported values are: {VALID_HEMISPHERES!s}"
-        )
-    return hemispheres
-
-
-def _check_view_is_valid(view) -> bool:
-    """Check whether a single view is one of two valid input types.
-
-    Parameters
-    ----------
-    view : :obj:`str` in {"anterior", "posterior", "medial", "lateral",
-        "dorsal", "ventral" or pair of floats (elev, azim).
-
-    Returns
-    -------
-    valid : True if view is valid, False otherwise.
-    """
-    if isinstance(view, str) and (view in VALID_VIEWS):
-        return True
-    return (
-        isinstance(view, Sequence)
-        and len(view) == 2
-        and all(isinstance(x, (int, float)) for x in view)
-    )
-
-
-def _check_views(views) -> list:
-    """Check whether the views passed to in plot_img_on_surf are correct.
-
-    Parameters
-    ----------
-    views : :obj:`list`
-        Any combination of strings in {"anterior", "posterior", "medial",
-        "lateral", "dorsal", "ventral"} and / or pair of floats (elev, azim).
-
-    Returns
-    -------
-    views : :obj:`list`
-        Views given as inputs.
-    """
-    invalid_views = [not _check_view_is_valid(view) for view in views]
-
-    if any(invalid_views):
-        raise ValueError(
-            "Invalid view definition!\n"
-            f"Got: {np.array(views)[invalid_views]!s}\n"
-            f"Supported values are: {VALID_VIEWS!s}"
-            " or a sequence of length 2"
-            " setting the elevation and azimut of the camera."
-        )
-
-    return views
 
 
 def _colorbar_from_array(
